backend/requirements.txt
# Backend Dependencies - FastAPI Server
# Core API framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# Resume parsing
PyMuPDF>=1.26.4
pdfplumber>=0.11.7
python-docx>=1.2.0
docx2txt>=0.9

# NLP and text processing
spacy>=3.8.7
nltk>=3.9.1
scikit-learn>=1.7.2
numpy>=2.3.3
pandas>=2.2.2

# AI and LLM
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.3.7
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Database
sqlalchemy>=2.0.20

# Utilities
python-dotenv==1.0.0
pydantic>=2.5.0
jinja2==3.1.2